// packages/datadog-api-client-v2/models/RoleRelationships.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RelationshipToPermissions } from './RelationshipToPermissions';
import { RelationshipToUsers } from './RelationshipToUsers';
/**
* Relationships of the role object.
*/
export declare class RoleRelationships {
    'permissions'?: RelationshipToPermissions;
    'users'?: RelationshipToUsers;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): RoleRelationships;
    static serialize(data: RoleRelationships): {
        [key: string]: any;
    };
    constructor();
}